refactor: Remove CoreData dependency
